Improve compute source hash
We need to ignore .git and content defined by .gitignore.
