+++ qubesbuilder/plugins/build_rpm/__init__.py
@@ -141,7 +141,7 @@ class RPMBuildPlugin(BuildPlugin, RPMDistributionPlugin):
 
             # Compare previous artifacts hash with current source hash
             if all(
-                self.get_source_hash()
+                self.component.get_source_hash()
                 == self.get_artifacts_info(stage, spec.with_suffix("").name).get(
                     "source-hash", None
                 )
@@ -265,6 +265,6 @@ class RPMBuildPlugin(BuildPlugin, RPMDistributionPlugin):
                 info = {
                     "srpm": source_info["srpm"],
                     "rpms": packages_list,
-                    "source-hash": self.get_source_hash(),
+                    "source-hash": self.component.get_source_hash(),
                 }
                 self.save_artifacts_info(stage=stage, basename=spec_bn, info=info)
